refactor: extract reloadCustomCommands from app.tsx
Move custom command loading logic (~30 lines) into src/app/commands-init.ts
and deduplicate the inline init-effect copy.

App keeps a thin useCallback wrapper. Removes unused loadCustomCommands
import. Zero behavioral change.

Commit-Queue-Session: cq_20260508160558_dad111b7260fef0bcd71f2b9
Coding-Agent: pi
Coding-Agent-Session: pi-019e06b6-830e-72ca-8378-e2ba2d3209c0
